// src/types/AssetTypes.ts
/**
 * Stable keys for project-owned visual assets.
 */
export enum AssetKey {
  BackgroundVillage = 'background-village',
  EnemyBig = 'enemy-big',
  EnemyNormal = 'enemy-normal',
  EnemySpike = 'enemy-spike',
  HouseVietnam = 'house-vietnam',
  PlayerCow = 'player-cow',
  PlayerDuck = 'player-duck',
  PlayerPeanut = 'player-peanut',
  WeaponFlipFlop = 'weapon-flipflop',
}